+++ test/generate.js
@@ -5,7 +5,8 @@ export const generateFile = ({
   ext,
   path = '/',
   dir_id = 'io.cozy.files.root-dir',
-  updated_at = ''
+  updated_at = '',
+  encrypted = false
 } = {}) => {
   let extension = ext
   if (extension === undefined) {
@@ -37,6 +38,7 @@ export const generateFile = ({
     path: `${path === '/' ? '' : path}/${prefix}${i}${extension}`,
     type,
     _type: 'io.cozy.files',
+    encrypted,
     ...optional
   }
 }